+++ database_ultimate_backup_lite/__manifest__.py
@@ -51,7 +51,7 @@
     'support': "reneluishs@gmail.com",
 
     'category': 'Administration',
-    'version': '16.0.1.1.0',
+    'version': '16.0.1.1.1',
 
     # Module dependencies
     'depends': ['base'],